requirements.txt
pandas>=1.5.0
boto3>=1.26.0
requests>=2.28.0
PyYAML>=6.0
urllib3>=1.26.0
python-dateutil>=2.8.0 